refactor(cli): hook dispatch helpers を cli/hook/mod.rs へ移行 (SPEC-1942)
SPEC-1942 SC-025 進捗継続。 cli.rs に残っていた hook 系 dispatch helper を
cli/hook/mod.rs に集約し、 hook family の自己完結性を強化。

cli/hook/mod.rs へ移動:
- run_hook (in-process hook dispatch entry)
- run_daemon_hook (internal daemon-hook fallback path)
- daemon_hook_argv (test-only argv builder、 pub(crate))
- write_internal_command_output (test-only stream relay、 pub(crate))
- prepare_daemon_front_door_for_path (daemon bootstrap、 pub)
- refresh_managed_assets_for_hook_front_door (worktree managed asset refresh、
  pub(crate) for cli.rs tests)

caller 更新:
- cli.rs run() dispatch: run_hook → hook::run_hook, run_daemon_hook → hook::run_daemon_hook
- main.rs (gwt binary): cli::prepare_daemon_front_door_for_path →
  cli::hook::prepare_daemon_front_door_for_path
- cli.rs tests: bare 呼び出しを crate::cli::hook::* qualified path に更新

cli.rs 1752 → 1545 行 (-207、 累計 -1156 from 2701、 -42.8%)。
hook/mod.rs 129 → 351 行 (+222)。 機能・契約・test 結果は不変
(lib 345 / 全 integration tests 全 pass、 clippy 0 / fmt clean)。
